\name{Cartesian2DCoordinateTransformation-class}
\Rdversion{1.1}
\docType{class}
\alias{Cartesian2DCoordinateTransformation-class}
\alias{show,Cartesian2DCoordinateTransformation-method}
\alias{numericOrNULL-class}
\title{Class "Cartesian2DCoordinateTransformation"}
\description{
    Virtual base class for cartesian 2D coordinate transformations such as 
    similarity and affine. Please read the section 'See Also' for further 
    documentation on slots and methods.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from 
    it.}
\author{German Carrillo}
\seealso{\code{\link{SimilarityTransformation-class}}, 
    \code{\link{AffineTransformation-class}}}
\keyword{classes}
