\name{control.points}
\docType{data}
\alias{control.points}
\title{Sample control points data.frame}
\description{
    This data.frame contains control points for defining a 2D cartesian 
    coordinate transformation. The control points are in both the source and 
    target coordinates.
}
\usage{data(control.points)}
\format{A data.frame consisting on 16 control points with ID, coordinates X and 
Y from the source dataset and coordinates X and Y from the target dataset.}
\keyword{datasets}
